+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## unreleased
 
+- add function to directly parse a module instead of a script
 - add a basic optimization mode
 - add a profiling mode
 - improve compilation time by using menhir with table mode